pbbbt

+++ Statistics.cs
@@ -48,8 +48,8 @@ namespace Taskmaster
 		public static long TouchCount { get; set; } = 0;
 		public static long TouchIgnore { get; set; } = 0;
 
-		public static double TouchTimeLongest { get; set; } = 0d;
-		public static double TouchTimeShortest { get; set; } = double.MaxValue;
+		public static long TouchTimeLongest { get; set; } = 0L;
+		public static long TouchTimeShortest { get; set; } = long.MaxValue;
 
 		public static int FatalErrors { get; set; } = 0;
